Modularised components and made it not work
