+++ CCTRI/CMakeLists.txt
@@ -1,9 +1,13 @@
 cmake_minimum_required(VERSION 3.10)
 project(TRIRG LANGUAGES CXX VERSION 0.1)
-add_executable(TRIRG TRIRG.cpp)
+add_executable(TRIRG TRIRG.cpp generalFunctions.cpp)
 add_library(randNums randNumGen.h)
 set_target_properties(randNums PROPERTIES LINKER_LANGUAGE CXX)
-target_link_libraries(TRIRG PUBLIC randNums)
+add_library(histogramOperations histogramOperations.h)
+set_target_properties(histogramOperations PROPERTIES LINKER_LANGUAGE CXX)
+add_library(RGStep RGStep.h)
+set_target_properties(RGStep PROPERTIES LINKER_LANGUAGE CXX)
+target_link_libraries(TRIRG PUBLIC randNums RGStep histogramOperations)
 set(CMAKE_CXX_STANDARD 11)
 set(CMAKE_CXX_STANDARD_REQUIRED true)
 set(SEED 12)
