+++ CCTRI/TRIRG.cpp
@@ -1,15 +1,16 @@
 #include "TRIRGConfig.h"
 #include "randNumGen.h"
+#include "histogramOperations.h"
+#include "RGStep.h"
 #include <iostream>
+
+// Conditional clause importing Eigen from either usr/include or from local
 #if __has_include("<Eigen/Dense>")
 # include <Eigen/Dense>
 #elif __has_include("../Eigen/eigen-master/Eigen/Dense")
 #include "../Eigen/eigen-master/Eigen/Dense"
 #endif
-//global installation of Eigen
-//#include <Eigen/Dense>
-//local installation of Eigen
-#include "../Eigen/eigen-master/Eigen/Dense"
+
 #include <cmath>
 #include <complex>
 #include <random>
@@ -20,6 +21,8 @@
 #include <unistd.h>
 using namespace std::chrono;
 
+
+
 using std::sin;
 using std::cos;
 using std::exp;
@@ -29,11 +32,10 @@ using std::acos;
 using Eigen::MatrixXd;
 using Eigen::Matrix;
 using Eigen::VectorXcd;
-const std::complex<double> i(0.0,1.0);
-const double twopi = acos(0.0) * 4;
+//const std::complex<double> i(0.0,1.0);
+//const double twopi = acos(0.0) * 4;
 const std::string filename = "TRIRG";
 randNums RNG;
-
 
 /*
 
@@ -56,192 +58,7 @@ std::string getPath()
   std::string fullpath = std::string( result, (count > 0) ? count : 0 );
   return fullpath.substr(0,fullpath.size()-filename.size());
 }
- 
-/*
----------------------------
-matrixReturnTRI
----------------------------
 
-usage:
-    given initial parameters, returns the matrix representing a TRI RG unit for those parameters
-
-parameters:
-    values for node spin-mixing angles p1 through 5 (0<=px<2pi)
-    normal scattering angles t  (0<=t[x]<2pi)
-    8 phases x representing phase accrued over one of the 8 edges in the RG node
-
-returns:
-    20x20 (consisting of complex doubles) matrix which represents the TRI RG cell entirely
-
-side effects:
-    none
-
-*/
-Matrix<std::complex<double>,20,20> matrixReturnTRI(vector<double> p, vector<double> t, vector<double> x){
-        Matrix<std::complex<double>,20,20> r {
-                {1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -cos(p[0])*sin(t[0])*exp(-i*x[0]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -sin(p[0])*exp(i*x[1]), 0.0, 0.0},
-                {0.0, 1.0, 0.0, 0.0, 0.0, -i*cos(p[0]) *cos(t[0]) *exp(i*x[2]), 0.0, 0.0, sin(p[0])*exp(-i*x[0]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -cos(p[0])*sin(t[0])*exp(i*x[1]), 0.0, 0.0},
-                {0.0, 0.0, 1.0, 0.0, 0.0, -cos(p[0])*sin(t[0]) * exp(i*x[2]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -i * cos(p[0]) * cos(t[0]) * exp(i*x[1]), 0.0, 0.0},
-                {0.0, 0.0, 0.0, 1.0, 0.0, sin(p[0]) * exp(i *x[2]), 0.0, 0.0, -i*cos(p[0]) * cos(t[0]) * exp(-i*x[0]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0},
- 
-                {-i* cos(p[1]) *cos(t[1]) *exp(-i*x[2]), 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0,  0.0, -sin(p[1])*exp(i *x[3]), 0.0, -cos(p[1]) *sin(t[1])* exp(-i *x[4]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0},
-                {0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, -cos(p[1])* sin(t[1])* exp(i*x[3]), 0.0, sin(p[1]) *exp(-i*x[4]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0},
-                {-sin(p[1]) *exp(-i*x[2]), 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0,  0.0, -i *cos(p[1]) *cos(t[1])* exp(i* x[3]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0},
-                {-cos(p[1]) *sin(t[1]) *exp(-i*x[2]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, -i *cos(p[1])* cos(t[1]) *exp(-i *x[4]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0},
- 
-                {0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -cos(p[2])* sin(t[2])* exp(-i*x[3]), 1.0, 0.0, 0.0, 0.0,  0.0, -sin(p[2])* exp(i*x[5]), 0.0, 0.0, -i* cos(p[2])* cos(t[2])* exp(-i *x[7]), 0.0, 0.0, 0.0},
-                {0.0, 0.0, -i *cos(p[2]) *cos(t[2]) *exp(i*x[0]), 0.0, 0.0, 0.0, 0.0, sin(p[2])* exp(-i*x[3]), 0.0, 1.0, 0.0, 0.0, 0.0, -cos(p[2]) *sin(t[2]) *exp(i*x[5]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0},
-                {0.0, 0.0, -cos(p[2])* sin(t[2]) *exp(i*x[0]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0,  0.0, -i *cos(p[2]) *cos(t[2]) *exp(i*x[5]), 0.0, 0.0, -sin(p[2])* exp(-i*x[7]), 0.0, 0.0,  0.0},
-                {0.0, 0.0, sin(p[2]) *exp(i*x[0]), 0.0, 0.0, 0.0,  0.0, -i *cos(p[2])* cos(t[2])* exp(-i *x[3]), 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, -cos(p[2])* sin(t[2])* exp(-i*x[7]), 0.0, 0.0, 0.0},
- 
-                {0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -i *cos(p[3]) *cos(t[3]) *exp(-i*x[5]), 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, -sin(p[3])* exp(i*x[6]), 0.0},
-                {0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -i*cos(p[3])* cos(t[3])* exp(i*x[4]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, -cos(p[3]) *sin(t[3])* exp(i*x[6]), 0.0},
-                {0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -cos(p[3])* sin(t[3])* exp(i*x[4]), 0.0, 0.0, 0.0, 0.0, -sin(p[3])* exp(-i*x[5]), 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, -i *cos(p[3])* cos(t[3]) *exp(i*x[6]), 0.0},
-                {0.0, 0.0, 0.0, 0.0, 0.0, 0.0, sin(p[3]) *exp(i*x[4]), 0.0, 0.0, 0.0, 0.0, -cos(p[3]) *sin(t[3]) *exp(-i*x[5]), 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0},
- 
-                {0.0, 0.0, 0.0, -i *cos(p[4])* cos(t[4]) *exp(-i*x[1]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -cos(p[4]) *sin(t[4])* exp(-i*x[6]), 1.0, 0.0, 0.0, 0.0},
-                {0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -i *cos(p[4]) *cos(t[4]) *exp(i*x[7]), 0.0, 0.0, 0.0, 0.0, 0.0, sin(p[4])* exp(-i*x[6]), 0.0, 1.0, 0.0, 0.0},
-                {0.0, 0.0, 0.0, -sin(p[4]) *exp(-i*x[1]), 0.0, 0.0, 0.0, 0.0,  0.0, -cos(p[4]) *sin(t[4]) *exp(i*x[7]), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0},
-                {0.0, 0.0, 0.0, -cos(p[4])* sin(t[4]) *exp(-i*x[1]), 0.0, 0.0, 0.0, 0.0, 0.0, sin(p[4]) *exp(i*x[7]), 0.0, 0.0, 0.0, 0.0, 0.0, -i *cos(p[4])* cos(t[4])* exp(-i*x[6]), 0.0, 0.0, 0.0, 1.0},
-        };
-        return r;
-
-
-}
-/*
----------------------------
-inputVectorReturnTRI
----------------------------
-
-usage:
-    given initial parameters, returns the vector representing the input channels of the TRI unit cell
-
-parameters:
-    values for node spin-mixing angle p (0<=p[x]<2pi)
-    normal scattering angles t (0<=t[x]<2pi)
-    the 4 possible inpout channels (0<=inputs[x]<1)
-
-returns:
-    1x20 (consisting of complex doubles) matrix/vector representing the inputs of the TRI unit cell
-*/
-Matrix<std::complex<double>,20,1> inputVectorReturnTRI(vector<double> p,vector<double> t,vector<double> inputs){
-    Matrix<std::complex<double>,20,1> result {
-        i* cos(p[0]) *cos(t[0]) *inputs[0], 0, sin(p[0]) *inputs[0],cos(p[0])* sin(t[0]) *inputs[0],
-         0, i*cos(p[1]) *cos(t[1])* inputs[1],cos(p[1])* sin(t[1])*inputs[1], -sin(p[1])* inputs[1],
-         0, 0, 0, 0,
-        cos(p[3])*sin(t[3]) *inputs[2], -sin(p[3])*inputs[2], 0, i*cos(p[3]) *cos(t[3])*inputs[2],
-        sin(p[4])*inputs[3], cos(p[4])* sin(t[4])* inputs[3], i * cos(p[4])* cos(t[4])*inputs[3], 0};
-    return result;
-}
-
-
-
-
-
-/*
----------------------------
-launder
----------------------------
-
-usage:
-    given a histogram, it generates a distribution according to that histogram (I call it launder because it is like laundering the points in the distribution)
-
-parameters:
-    histPoints: array of histogram points
-    min: minimum value that can be attained in the distribution
-    max: maximum value that can be attained in the distribution
-    length: how many values to generate
-
-returns:
-    a vector of specified length consisting of values that conform to the histogram provided
-
-*/
-vector<double> launder(vector<int> histPoints, double min, double max, int length){
-    // determine how wide each bin is
-    double binWidth = (max-min)/histPoints.size();
-    double histPointsSum = std::reduce(histPoints.begin(),histPoints.end());
-    // make the histogram points normed by dividing each element by total * binWidth
-    for(double i : histPoints){
-        i = i / (histPointsSum * binWidth);
-    }
-    // find maximum value from the histogram to create a cutoff for generating values
-    double hmax  = *std::max_element(histPoints.begin(),histPoints.end());
-    // initialise laundered array
-    vector<double> laundered(length);
-    // populate laundered array
-    for(int i{0};i<length;i++){
-        // generate a 'prospect point' which could potentially be within the distribution
-        vector<double> prospectPoint = {RNG.randDouble(min,max),RNG.randDouble(0,hmax)};
-        // determine which bin this prospect number should fall into
-        int binNo = std::floor((prospectPoint[0]-min)/binWidth);
-        // keep generating prospect points until one falls within the distribution
-        while(histPoints[binNo] < prospectPoint[1]){
-            prospectPoint = RNG.randDouble(min,max,2);
-            binNo = std::floor((prospectPoint[0]-min)/binWidth);
-        }
-        // add the point to the distribution
-        laundered[i] = prospectPoint[0];
-
-    }
-    return laundered;
-}
-
-/*
----------------------------
-binCounts
----------------------------
-
-usage:
-    given a set of datapoints, place them into bins of set width, returning the datapoints of the histogram representing the data
-
-parameters:
-    data: an array of data points to bin
-    min: minimum value of the range of the data points
-    max: maximum value of the range of the data points
-    binWidth: desired width of bins
-
-returns:
-    int array
-
-*/
-vector<int> binCounts(double* data, double min, double max, double binWidth, int length){
-    int amountOfBins = (int)std::ceil((max-min)/binWidth);
-    vector<int> bins(amountOfBins);
-    for(int i{0};i<length;i++){
-        int binNo = (int)std::floor((data[i]-min)/binWidth);
-        bins[binNo]+=1;
-    }
-    return bins;
-}
-
-
-/*
----------------------------
-renormalise
----------------------------
-
-usage:
-    performs an analytic renormalisation transformation based on an RSRG unit cell
-
-parameters:
-    angleVector: double array of spin mixing angles
-    scatteringAngleVector: double array of scattering angles (thetas)
-    phases: double array of random phases between nodes of the unit cell
-    inputs: double array of input values into the system
-
-returns:
-    a single double representing the renormalised scattering angle for the whole unit cell
-
-*/
-double renormalise(vector<double> angleVector, vector<double> scatteringAngleVector, vector<double> phases, vector<double> inputs){
-    Matrix<std::complex<double>,20,20> system = matrixReturnTRI(angleVector,scatteringAngleVector,phases);
-    Matrix<std::complex<double>,20,1> inputvec = inputVectorReturnTRI(angleVector,scatteringAngleVector,inputs);
-    Matrix<std::complex<double>,20,20> inv = system.inverse();
-    Matrix<std::complex<double>,20,1> tmp = inv*inputvec;
-    double tval = std::asin(std::abs(tmp(19))/(cos(std::asin(std::sqrt(std::pow(std::abs(tmp(14)),2) + std::pow(std::abs(tmp(1)),2))))));
-    return tval;
-
-}
 
 int main(int argc, char* argv[])
 {
