+++ CCTRI/histogramOperations.h
@@ -1,0 +1,14 @@
+#include <random>
+#include <cmath>
+#include <complex>
+
+using std::vector;
+
+#ifndef FUNCTIONS_H_INCLUDED
+#define FUNCTIONS_H_INCLUDED
+vector<double> launder(vector<int> histPoints, double min, double max, int length);
+
+vector<int> binCounts(double* data, double min, double max, double binWidth, int length);
+double renormalise(vector<double> angleVector, vector<double> scatteringAngleVector, vector<double> phases, vector<double> inputs);
+
+#endif
